# service: Use sync pool for Sign/Verify request headers